requirements.txt
googlesearch-python
beautifulsoup4
requests
colorama